refactor: stabilize review reopen flow and split discord runtime
