+++ src/server/routes/dispatches/mod.rs
@@ -20,6 +20,7 @@ pub(crate) use outbox::{dispatch_outbox_loop, queue_dispatch_followup};
 
 // ── Re-exports: Thread reuse ─────────────────────────────────
 pub(super) use thread_reuse::clear_all_threads;
+pub(crate) use thread_reuse::validate_channel_thread_maps_on_startup;
 pub use thread_reuse::{get_card_thread, get_pending_dispatch_for_thread, link_dispatch_thread};
 
 // ── Shared utilities (used by both discord_delivery and thread_reuse) ──
