+++ src/server/routes/dispatches/thread_reuse.rs
@@ -143,6 +143,220 @@ pub(in crate::server::routes) fn clear_all_threads(conn: &rusqlite::Connection, 
     .ok();
 }
 
+pub(crate) async fn validate_channel_thread_maps_on_startup(
+    db: &crate::db::Db,
+    token: &str,
+) -> (usize, usize) {
+    let client = match reqwest::Client::builder()
+        .timeout(std::time::Duration::from_secs(2))
+        .build()
+    {
+        Ok(client) => client,
+        Err(e) => {
+            tracing::warn!("[dispatch] failed to build startup thread-map validator client: {e}");
+            return (0, 0);
+        }
+    };
+
+    validate_channel_thread_maps_on_startup_with_base_url(
+        db,
+        &client,
+        token,
+        "https://discord.com/api/v10",
+    )
+    .await
+}
+
+async fn validate_channel_thread_maps_on_startup_with_base_url(
+    db: &crate::db::Db,
+    client: &reqwest::Client,
+    token: &str,
+    base_url: &str,
+) -> (usize, usize) {
+    let rows: Vec<(String, String, Option<String>)> = match db.lock() {
+        Ok(conn) => conn
+            .prepare(
+                "SELECT id, channel_thread_map, active_thread_id
+                 FROM kanban_cards
+                 WHERE channel_thread_map IS NOT NULL
+                   AND TRIM(channel_thread_map) != ''
+                   AND TRIM(channel_thread_map) != '{}'",
+            )
+            .and_then(|mut stmt| {
+                stmt.query_map([], |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?)))
+                    .map(|rows| rows.filter_map(|row| row.ok()).collect())
+            })
+            .unwrap_or_default(),
+        Err(e) => {
+            tracing::warn!("[dispatch] startup thread-map validation skipped (db lock): {e}");
+            return (0, 0);
+        }
+    };
+
+    let mut checked = 0usize;
+    let mut cleared = 0usize;
+
+    for (card_id, map_json, active_thread_id) in rows {
+        let Ok(mut map) =
+            serde_json::from_str::<serde_json::Map<String, serde_json::Value>>(&map_json)
+        else {
+            if let Ok(conn) = db.lock() {
+                conn.execute(
+                    "UPDATE kanban_cards SET channel_thread_map = NULL WHERE id = ?1",
+                    [&card_id],
+                )
+                .ok();
+            }
+            cleared += 1;
+            continue;
+        };
+
+        let mut changed = false;
+        let mut removed_active = false;
+        let snapshot: Vec<(String, Option<String>)> = map
+            .iter()
+            .map(|(channel_id, thread_id)| {
+                (
+                    channel_id.clone(),
+                    thread_id.as_str().map(std::string::ToString::to_string),
+                )
+            })
+            .collect();
+
+        for (channel_id_raw, thread_id) in snapshot {
+            checked += 1;
+            let Some(thread_id) = thread_id else {
+                map.remove(&channel_id_raw);
+                changed = true;
+                cleared += 1;
+                continue;
+            };
+            let Ok(expected_parent) = channel_id_raw.parse::<u64>() else {
+                map.remove(&channel_id_raw);
+                changed = true;
+                cleared += 1;
+                if active_thread_id.as_deref() == Some(thread_id.as_str()) {
+                    removed_active = true;
+                }
+                continue;
+            };
+
+            let thread_info_url =
+                format!("{}/channels/{}", base_url.trim_end_matches('/'), thread_id);
+            let response = match client
+                .get(&thread_info_url)
+                .header("Authorization", format!("Bot {}", token))
+                .send()
+                .await
+            {
+                Ok(response) => response,
+                Err(e) => {
+                    tracing::info!(
+                        "[dispatch] startup thread-map validation clearing {} -> {} after request error: {}",
+                        card_id,
+                        thread_id,
+                        e
+                    );
+                    map.remove(&channel_id_raw);
+                    changed = true;
+                    cleared += 1;
+                    if active_thread_id.as_deref() == Some(thread_id.as_str()) {
+                        removed_active = true;
+                    }
+                    continue;
+                }
+            };
+
+            if !response.status().is_success() {
+                tracing::info!(
+                    "[dispatch] startup thread-map validation clearing {} -> {} after {}",
+                    card_id,
+                    thread_id,
+                    response.status()
+                );
+                map.remove(&channel_id_raw);
+                changed = true;
+                cleared += 1;
+                if active_thread_id.as_deref() == Some(thread_id.as_str()) {
+                    removed_active = true;
+                }
+                continue;
+            }
+
+            let body: serde_json::Value = match response.json().await {
+                Ok(body) => body,
+                Err(e) => {
+                    tracing::info!(
+                        "[dispatch] startup thread-map validation clearing {} -> {} after json error: {}",
+                        card_id,
+                        thread_id,
+                        e
+                    );
+                    map.remove(&channel_id_raw);
+                    changed = true;
+                    cleared += 1;
+                    if active_thread_id.as_deref() == Some(thread_id.as_str()) {
+                        removed_active = true;
+                    }
+                    continue;
+                }
+            };
+
+            let parent_id = body
+                .get("parent_id")
+                .and_then(|value| value.as_str())
+                .and_then(|value| value.parse::<u64>().ok())
+                .unwrap_or_default();
+
+            if parent_id != expected_parent {
+                tracing::info!(
+                    "[dispatch] startup thread-map validation clearing {} -> {} (parent {} != {})",
+                    card_id,
+                    thread_id,
+                    parent_id,
+                    expected_parent
+                );
+                map.remove(&channel_id_raw);
+                changed = true;
+                cleared += 1;
+                if active_thread_id.as_deref() == Some(thread_id.as_str()) {
+                    removed_active = true;
+                }
+            }
+        }
+
+        if !changed {
+            continue;
+        }
+
+        let new_map = if map.is_empty() {
+            None
+        } else {
+            Some(serde_json::to_string(&map).unwrap_or_else(|_| "{}".to_string()))
+        };
+        let new_active_thread_id = if removed_active {
+            map.values()
+                .find_map(|value| value.as_str())
+                .map(std::string::ToString::to_string)
+        } else {
+            active_thread_id.clone()
+        };
+
+        if let Ok(conn) = db.lock() {
+            conn.execute(
+                "UPDATE kanban_cards
+                 SET channel_thread_map = ?1,
+                     active_thread_id = ?2
+                 WHERE id = ?3",
+                rusqlite::params![new_map, new_active_thread_id, card_id],
+            )
+            .ok();
+        }
+    }
+
+    (checked, cleared)
+}
+
 /// Try to reuse an existing Discord thread for a dispatch.
 /// Returns `Some(true)` if reuse succeeded, `Some(false)` if the thread exists but is locked,
 /// or `None` if the thread couldn't be accessed (deleted, wrong parent, etc.).
@@ -328,6 +542,101 @@ pub(super) async fn try_reuse_thread(
     } else {
         tracing::warn!("[dispatch] Failed to send message to reused thread {thread_id}");
         None
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use axum::{
+        Json, Router, extract::Path, http::StatusCode, response::IntoResponse, routing::get,
+    };
+    use serde_json::json;
+
+    fn test_db() -> crate::db::Db {
+        let conn = rusqlite::Connection::open_in_memory().unwrap();
+        conn.execute_batch("PRAGMA foreign_keys=ON;").unwrap();
+        crate::db::schema::migrate(&conn).unwrap();
+        crate::db::wrap_conn(conn)
+    }
+
+    async fn spawn_thread_info_server() -> (String, tokio::task::JoinHandle<()>) {
+        async fn channel(Path(thread_id): Path<String>) -> impl IntoResponse {
+            match thread_id.as_str() {
+                "thread-valid" => (
+                    StatusCode::OK,
+                    Json(json!({"id":"thread-valid","parent_id":"111"})),
+                )
+                    .into_response(),
+                "thread-wrong-parent" => (
+                    StatusCode::OK,
+                    Json(json!({"id":"thread-wrong-parent","parent_id":"999"})),
+                )
+                    .into_response(),
+                _ => StatusCode::NOT_FOUND.into_response(),
+            }
+        }
+
+        let app = Router::new().route("/channels/{thread_id}", get(channel));
+        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
+        let addr = listener.local_addr().unwrap();
+        let handle = tokio::spawn(async move {
+            axum::serve(listener, app).await.unwrap();
+        });
+        (format!("http://{addr}"), handle)
+    }
+
+    #[tokio::test]
+    async fn startup_validation_clears_missing_and_mismatched_thread_bindings() {
+        let db = test_db();
+        {
+            let conn = db.lock().unwrap();
+            conn.execute(
+                "INSERT INTO kanban_cards (
+                    id, title, status, priority, channel_thread_map, active_thread_id,
+                    created_at, updated_at
+                ) VALUES (
+                    'card-thread-map-startup', 'Issue #335', 'review', 'medium',
+                    '{\"111\":\"thread-valid\",\"222\":\"thread-missing\",\"333\":\"thread-wrong-parent\"}',
+                    'thread-missing',
+                    datetime('now'), datetime('now')
+                )",
+                [],
+            )
+            .unwrap();
+        }
+
+        let client = reqwest::Client::builder().build().unwrap();
+        let (base_url, server_handle) = spawn_thread_info_server().await;
+        let (checked, cleared) = validate_channel_thread_maps_on_startup_with_base_url(
+            &db,
+            &client,
+            "test-token",
+            &base_url,
+        )
+        .await;
+        server_handle.abort();
+
+        assert_eq!(checked, 3);
+        assert_eq!(cleared, 2);
+
+        let conn = db.lock().unwrap();
+        let (map_json, active_thread_id): (Option<String>, Option<String>) = conn
+            .query_row(
+                "SELECT channel_thread_map, active_thread_id
+                 FROM kanban_cards
+                 WHERE id = 'card-thread-map-startup'",
+                [],
+                |row| Ok((row.get(0)?, row.get(1)?)),
+            )
+            .unwrap();
+
+        let map: serde_json::Value =
+            serde_json::from_str(map_json.as_deref().unwrap_or("{}")).unwrap();
+        assert_eq!(map["111"], "thread-valid");
+        assert!(map.get("222").is_none());
+        assert!(map.get("333").is_none());
+        assert_eq!(active_thread_id.as_deref(), Some("thread-valid"));
     }
 }
